requirements.txt
fastapi
fastapi-cli
uvicorn
cassandra-driver==3.29.1
confluent-kafka==2.4.0
kafka-python==2.0.2
pydantic==2.7.2
docker==7.1.0
pandas==2.0.3
pytest
httpx
pytest-asyncio
python-dotenv
python-jose
passlib[bcrypt]